Fix tests after merge

+++ src/net/net_transport_test.go
@@ -43,8 +43,8 @@ func TestNetworkTransport_Sync(t *testing.T) {
 				Body: poset.WireBody{
 					Transactions:          [][]byte(nil),
 					SelfParentIndex:       1,
-					OtherParentCreatorIDs: []int{10},
-					OtherParentIndexes:    []int{0},
+					OtherParentCreatorIDs: []int64{10},
+					OtherParentIndexes:    []int64{0},
 					CreatorID:             9,
 				},
 			},
@@ -108,8 +108,8 @@ func TestNetworkTransport_EagerSync(t *testing.T) {
 				Body: poset.WireBody{
 					Transactions:          [][]byte(nil),
 					SelfParentIndex:       1,
-					OtherParentCreatorIDs: []int{10},
-					OtherParentIndexes:    []int{0},
+					OtherParentCreatorIDs: []int64{10},
+					OtherParentIndexes:    []int64{0},
 					CreatorID:             9,
 				},
 			},
@@ -180,8 +180,8 @@ func TestNetworkTransport_PooledConn(t *testing.T) {
 				Body: poset.WireBody{
 					Transactions:          [][]byte(nil),
 					SelfParentIndex:       1,
-					OtherParentCreatorIDs: []int{10},
-					OtherParentIndexes:    []int{0},
+					OtherParentCreatorIDs: []int64{10},
+					OtherParentIndexes:    []int64{0},
 					CreatorID:             9,
 				},
 			},
